[SYSTEMML-1044] Generalized spark mapleftindex (left/right broadcasts)
This patch generalizes our existing spark mapleftindex to allow not just
the rhs but now also the lhs (under special conditions) to be broadcast.
This can significantly reduce the memory pressure for sparse-dense left
indexing as the sparse left-hand-side matrix might actually be much
smaller in absolute size, although regarding dimensions, the lhs is
always larger than the rhs.

This fixes OOMs of our data generator for descriptive statistics
(SYSTEMML-1044).