[SYSTEMML-1369] New loop vectorization rewrite for indexed copies
As explained in https://issues.apache.org/jira/browse/SYSTEMML-1369,
this patch introduces an auto loop vectorization rewrite for indexed
copies. For example, we now rewrite the following loop

parfor (i in 1:ncol(labels))
   topics[id, i] = labels[1, i];
   
to a simple left indexing of topics[id, 1:ncol(labels)] = labels[1,
1:ncol(labels)]. This applies to for and parfor loops. Furthermore, this
patch also fixes size update issues of the existing loop vectorization
rewrites vectorizeElementwiseBinary and vectorizeElementwiseUnary.

On a scenario of an 1K x 1K dense topics matrix, an outer loop over id,
and regular for loops, the runtime improved from 33s (3438s without
update-in-place) to 0.2s. Note that on large out-of-core matrices, the
improvements are even larger because we cannot apply update-in-place
there.