[MINOR] Cleanup hop visit status and its usage, remove hop print
The hop visit status - used for memoization on dag traversal - was
originally multi-valued but later reduced to binary states. This patch
replaces the visit status with a simple boolean variable which
simplifies its usage. Furthermore, this patch also removes the hop
printme debug output as it's superseded by the hop explain.