[SYSTEMML-1317] Avoid unnecessary copies on creating combiner blocks
For most aggregations (e.g., merge of partial blocks, or aggregate by
key), we use combineByKey to allow for update-in-place when adding to
combiner blocks. However, so far we deep copy - often unnecessarily -
blocks when creating the combiner blocks. This creates unnecessary
memory pressure in scenarios where there is almost no aggregation. We
now exploit the knowledge of temporary blocks to allow aggregations with
shallow copy to reduce garbage collection overhead and memory pressure.