[SYSTEMML-1404] Robustness ultra-sparse cell to binary block converters
The current implementations of spark rdd converters for 'textcell to
binary block', 'binary cell to binary block', and 'coordinate to binary
block' have severe robustness issues for ultra-sparse inputs, in terms
of potential OOMs at the driver or executors. All these converters
perform a union of partial blocks with a set of empty blocks before
shuffle but the collection of empty blocks is simply parallelized from
the driver. 

Consider a scenario of an ultra-sparse 30M x 30M matrix with sparsity
10^-6, which is arguable the worst-case with a single non-zero per block
on average. The number of blocks is 900M and these billion objects
(pairs of matrix indexes and empty matrix blocks) are already 64GB,
leading to OOMs at the driver. Furthermore, the cell converters hold
partial blocks in MSCR format which is two orders of magnitude larger
than a cell representation leading to OOMs at the executors.

This patch solves both problems by (1) generating empty blocks in a
scalable manner (with awareness of the target size of output
partitions), and (2) converting temporary ultra-sparse partial blocks to
COO format (which does not have any overhead for row offsets and hence
is close to the cell representation).

Finally, this also includes a fix for caching of binary cell rdds, which
would have returned null under specific conditions.