[SYSTEMML-1314] Robust memory handling for RDD parallelization
So far we only checked on RDD parallelization if the pinned matrix block
as well as the partitioned matrix fit together into the driver memory
budget; otherwise the matrix is exported (w/ a multi-threaded write in
binary block format to HDFS). However, since these parallelized RDDs
continue to reside in the driver, subsequent CP operations, broadcasts,
collects, or RDD parallelization might run into OOMs at the driver. 

This patch establishes a memory manager for parallelized RDDs that
allows to reserve memory and keeps track of currently referenced RDDs.
We keep the traditional 70% operation memory and 15% bufferpool, but use
additional 10% for potentially parallelized RDDs. Note that this
slightly over-provisions available memory as usually only 85-90% of max
heap are available to the application, but it prevents unnecessarily
conservative memory handling for the common case. 

Down the road, however, we should unify the entire buffer pool and
memory handling to manage CP intermediates, broadcasts, parallelized
RDDs, etc in one common memory manager.