[SYSTEMML-1315] Robustness spark mapmm (handle large output partitions)
Our spark mapmm (broadcast-based matrix multiply) instruction already
repartitions the input RDD for shapes that potentially create large
outputs in order to control the size of output partitions and with that
avoid failures due to Spark's 2GB limitation per partition. However,
with out perftest-stratstats script we encountered special cases of
dense-sparse matrix multiplications, where this was not enough and we
ran unnecessarily into OOMs. In detail, the sparse matrix was the
broadcast (because it was smaller) but the dense input had a small
number of blocks which limited the maximum number of partitions. 

This patch extends the mapmm instruction to consider these cases and
potentially flip broadcast and rdd at runtime if this is beneficial for
controlling the output sizes.