[SYSTEMML-1423] Fix OOMs ultra-sparse spark/mr rand (stream nnz/block)
Our rand operations of all backends materialize the number of non-zeros
per block. For ultra-sparse matrices, this can - even for distributed
operations - quickly lead to out-of-memory situations at the driver. We
now create these nnz per block as a stream that can be lazily consumed.
This fixes OOMs when generating ultra-sparse matrices with spark or mr
instructions. CP operations still materialize this stream in order to
allow repeated access for multi-threaded computation, which is fine
because this intermediate memory requirement is already reflected in the
hops memory estimate.

Furthermore, this patch also fixes an issue with the generation of nnz
per block for matrices with nnz < numBlocks. So far, the number of
blocks with non-zeros was drawn as a single random number which could
create heavily skewed data. We now simply select blocks with non-zeros
using a probability of nnz/numblocks.