[SYSTEMML-1310] New parfor block partitioning (compiler/runtime/tests)
This patch extends the parfor optimizer and runtime with the support for
block partitioning, i.e., the partitioning and access of batches of rows
or columns. Batches are required to be aligned and smaller or equal than
the blocksize in the partitioning dimension, but can be arbitrarily
large in the other dimension. So far, only the remote spark data
partitioner supports these new partitioning formats but the resulting
partitions can be used either in local or remote parfor operators.