Adding a multi-GPU MNIST LeNet example.
In order to exploit a multi-GPU setup, this (1) sets the degree of
parallelism (hardcoded for now), (2) adjusts the inner `for` loop
(previously for iterations) to extract out groups of mini-batches,
where the number of mini-batches is equal to the degree of parallelism,
(3) adds an inner `parfor` to parallelize the computation of gradients
over a group of mini-batches (i.e. one mini-batch per GPU) and store
the gradients, and (4) updates the model after the `parfor` loop using
the aggregated (averaged) gradients.