[SYSTEMML-1291] Size bounding codegen plan and class caches (LRU, 1024)
The code generator uses a plan cache to reuse generated classes across
DAGs and during DAG recompilation in order to significantly reduce javac
and JIT compilation overheads. However, so far this plan cache was
unbounded in size, which can lead to various out-of-memory situations as
the garbage collector cannot unload unused classes. This patch
establishes a size-bounded plan cache, which maintains usage order and
evicts the least-recently-used (LRU) plan and corresponding entries from
the class cache, if the maximum number of plans (1024) is exceeded. 

Note that we use a hand-crafted eviction policy instead of soft
references to maintain consistency across plan and class caches and to
allow for better debugging.