[SYSTEMML-1418,1293,1424] New cost-based codegen plan selection
This patch introduces a new cost-based plan selector for code generation
including an initial cost model, which will become the default plan
selector in the future. The core ideas are (1) to determine partitions,
i.e, maximally connected sub graphs of memo table entries, (2) to
determine potential materialization points (w/ >=2 consumers) per
partition, and (3) to enumerate and cost alternative plans per partition
via a branch and bound algorithm. 

Initial end-to-end experiments on L2SVM, MLogreg, and GLM are very
promising as the resulting performance is equivalent or better than both
existing plan selection heuristics (fuse-all, fuse-no-redundancy).
However, this selector is not yet applied by default, because the cost
model needs further validation and the branch and bound pruning
conditions for skipping of subsets of plans are not integrated yet.