[SYSTEMML-1507] Improved compilation of rowwise codegen templates
This patch makes two small yet effective improvements to the
compilation of rowwise codegen templates:

1) Improved opening condition: We now open new row templates for each
matrix-colvector binary operations. The final decision on row vs cell
templates still is up to the cost-based plan selection but this extended
exploration open more opportunities.

2) Reduced number of temporary vector intermediates: We now apply basic
memoization on computing the temporary, thread-local memory requirements
in order to avoid double counting in complex DAG structures.

Note that this patch also includes a fix for size updates in spark spoof
instructions and modifies tests for cellwise and outer product templates
because the new opening condition now led to a row template on the
original expressions.