[SYSTEMML-1514] Fix codegen cost estimation (two-level memoization)
This patch fixes the cost estimation of fusion plans for complex fused
operators with internal DAG structures, where we mistakenly double
counted compute costs. Similarly, we incorrectly double counted costs of
materialized intermediates. The core idea is a two-level memoization,
i.e., memoization for pairs of hops and cost vectors which allows a
proper memoization but at the same time the evaluation of costs for
overlapping fused operators with redundant computation. 

Additionally, this patch also hardens the compilation of
multi-aggregates to ensure matching input dimensions and to exclude
partial rowwise fusion plans that cover matrix multiplications.