[SYSTEMML-1517] Improved memory handling compressed matrix-vector mult
This patch improves the temporary memory handling of compressed
matrix-vector and vector-matrix multiplication. So far, we repeatedly
allocated double and int vectors in the number of values per column
group, which causes unnecessary garbage collection overhead on openjdk.
We now allocate these vectors once for all column groups with the
maximum number of values and each column group simply resets and uses a
subset of this temporary memory.