[SYSTEMML-1547] Cache-conscious tsmm post-processing (copy upper, nnz)
This patch improves the performance of tsmm postprocessing including (1)
the copy of the upper triangular matrix to the lower triangle, and (2)
the number of non-zeros computation of the output. We now use a
cache-conscious scheme of copying similar to dense-dense transpose
including the fused maintenance of nnz. 

On a scenario of 10 iterations running outer products of 15K vectors
(with 15K x 15K outputs), this patch improved end-to-end execution time
from 16.2s to 11.4s