[SYSTEMML-1550] Improved cp dim/sparsity check (exploit worst-case est)
This patch improves the operator selection of all hops by taking the
worst-case estimates in account during checks for valid cp dimensions
and sparsity. With this improvement, we are able to avoid unnecessary
distributed operations for hops with large dimensions and unknown exact
but known worst-case sparsity estimates such as rexpand.