[SYSTEMML-1564] Add a Java test suite wrapper around `nn` DML test suite
The `nn` library contains it's own DML test suite for gradient checks
and unit tests that produces "ERROR: ..." messages if any of the
mathematical operations return incorrect results.  Note that this has
helped to find mathematical bugs in the library & engine that do not
result in JVM exceptions.  This commit simply adds a Java test harness
that wraps these DML tests, parsing the standard out for "ERROR:"
messages.  A new `scripts` test package has been created, with the aim
of eventually testing all of our DML algorithms.  Additionally, the
ParserWrapper has been updated to look in the `scripts` folder as a
fallback, which allows for easy testing of our actual scripts.