[SYSTEMML-618] Stabilizing the `two_layer_affine_l2_net` grad check.
The example `two_layer_affine_l2_net` network uses a ReLU nonlinearity,
could result in a false-negative in which the test fails due to a kink
being crossed in the ReLU nonlinearity.  This occurs when the tests,
f(x-h) and f(x+h), end up on opposite sides of the zero threshold of
max(0, fx).  This change aims to stabilize the test, while still
allowing for the usage of the ReLU nonlinearity.  Although we have
stabilized it, we will still remove it from automated testing to avoid
false failures.  In the future, we can explicitly check for this
scenario and rerun the test automatically.  For manual testing, simply
rerun the tests.