[SYSTEMML-1574] Fix serialized caching of ultra-sparse matrices
Previously, we stored all sparse matrices that are expected to exceed
aggregate memory in serialized form in order to eliminate unnecessary
in-meomry storage overheads. With the introduction of csr we disabled
this code path as csr's in-memory representation is as large as its
serialized form. However, this is not true for ultra-sparse matrices
which still benefit from serialized storage. Accordingly this patch
tweaks the corresponding selecting of checkpoint storage levels.