[SYSTEMML-1575] DataType Change Test Fix
While working on SYSTEMML-1554, an additional bug was uncovered.
Specifically, with the new IPA scalar replacement enhancement, the
`org.apache.sysml.test.integration.functions.misc.DataTypeChangeTest#testDataTypeChangeValidate4c`
test started to fail due to attempts to cast a Matrix to a Scalar
object during IPA scalar replacement.  This was due to a bug in which
if the calling program reassigned a variable to the output of a
function, and the datatype differed between that variable and the
function output, the calling program's variable map would not be
updated to remove this variable.  Thus, a variable of the incorrect
datatype would remain in the calling program, and  when the new IPA
scalar replacement ran over the rest of the calling program, it would
attempt to replace a scalar using the value in the calling program
variable map, which would incorrectly still be a matrix type.  This
bug did not show up before since we did not previously employ IPA
scalar replacement.

This fix updates the `extractFunctionCallReturnStatistics` method
to remove the variable from the calling program's variable map if
the calling program is reassigning a variable to the output of this
function and the datatypes differ.