[SYSTEMML-1595] Fix missing block sizes for mlcontext/jmlc outputs
This patch fixes the rewrite for exchanging persistent with transient
writes, which is used in the mlcontext and jmlc APIs. Since the original
persistent write might be an external format without blocksizes, we now
always set the blocksize of the new transient write (which always
requires block sizes) according to write input. Note that this did not
show up before, because the twrite usually just translates to a renaming
of variable names, which preserves the meta data of the passed
intermediate.