[SYSTEMML-1621] Fix value type inference scalar-matrix max/min/log/ppred
So far the value type inference for binary builtin functions simply
forwarded the value type of the left input which produced wrong meta
data for scalar-matrix operations where the scalar is of type integer
because the output matrix is always of type double. This patch fixes
this value type inference and consolidates the code path via a shared
primitive to set output properties of binary operations.