[SYSTEMML-1627] Fix guarded parallelize of matrices collected from rdds
This patch fixes special cases of rdd construction, where rdds are
created from in-memory matrices that have been previously collected from
rdds. Specifically, this targets guarded parallelize, which exports
matrices and creates the rdd from the hadoop file. So far we only
exported dirty in-memory matrices but not collected matrices (which are
not marked as dirty). Accordingly, following rdd operations fail with
file not found exceptions. This happens, for example, in special cases,
where unary operations (1 inputs, 1 output) run in CP, but binary
operations (2 outputs, 1 input) run in SPARK and we have to fall back to
guarded parallelize.