[SYSTEMML-1289] Codegen cell-wise operations over compressed matrices
This patch generalized the codegen cell-wise operations (no aggregation,
row aggregates, full aggregates) to support compressed matrix blocks as
main or side inputs, without full decompression. The approach is to use
column-wise access in the operator skeletons to make a single pass over
the compressed representation. As a basis for these operations, this
patch introduces a column group iterator and encoding-format-specific
iterators for convenient access to row partitions of compressed values
with or without zero values. These iterators can be reused for all other
codegen templates as well as special cases of decompression.