[SYSTEMML-1641] Fix rewrite 'matrix multiplication chain optimization'
A recent change, that simplified the hop visit status handling,
introduced a severe bug into the rewrite for matrix multiplication chain
optimization, which affectively disabled this rewrite. This issue has
huge performance implications for scripts that do not provide a good
initial matrix multiplication order. This patch simply fixes the issue,
and adds tests that explicitly check for the correct matrix product
order to avoid such an issue in the future.