[SYSTEMML-1644] Fix correctness dense-dense matrix mult w/ short lhs
This patch fixes an issue of incorrect results for special cases of
dense-dense matrix multiplication with short left-hand-side (1 < x <= 16
rows) and a right-hand-side of more than 1024 features. The underlying
issues were (1) incorrect indexing in the rhs for all but the first
column cache block of 1024 columns, and (2) wrong rhs row boundaries.
Hence, it affects, for example, Mlogreg and Kmeans with a number of
classes/ centroids <= 16 and a feature matrix with > 1024 features.