[SYSTEMML-1140] Performance bufferpool (shallow serialize for sparse)
This patch addresses performance issues due to serialization overhead of
sparse matrices (in MCSR format) on buffer pool write. The basic idea is
an extended shallow serialize for sparse matrices. Shallow serialize
(which simply keeps a strong reference instead of serializing the
matrix) was so far only used for dense matrices and sparse matrices in
CSR format because their in-memory size is equivalent to their
serialized size. For MCSR (our default sparse block), the in-memory
representation has some overhead, so serialization helps to avoid
unnecessary evictions to disk. However, as the number of columns (or nnz
per row) grows this overhead becomes negligible. Hence, we now use an
overhead threshold of 30% and use a shallow serialize whenever the
overhead is below this threshold.

For example, on a scenario of an 100K x 10K matrix with sparsity 0.1 and
20 iterations of X = X * i, this patch improved the end-to-end runtime
from 65s to 32.7s.