[SYSTEMML-1660] Performance sparse removeEmpty/order (shallow copy)
Similar to dense vector transpose and dense row-wise reshape, this patch
improves performance of sparse removeEmpty(rows) and order via shallow
copy of sparse rows, exploiting the fact that removeEmpty(rows) and
order do not modify the actual sparse rows. This shallow copy (by
reference) is safe due to copy-on-write semantics and safe
update-in-place initialization. Note that the recent change of shallow
serialize of sparse matrices on buffer pool write (SYSTEMML-1140),
enabled this optimization potential, because without shallow serialize
we would still effectively copy the entire matrix.

Experiments over different matrix shapes and sparsity showed huge
improvements of orders of magnitude. For example, here are the results
for 20 iterations of the respective single-node in-memory operations:

1) removeEmpty(rows), 20 iterations: 
1M x 1K, sparsity=0.1:     10.3s -> 1.4s
100K x 10K, sparsity=0.1:  7.1s -> 160ms
10K x 100K, sparsity=0.1:  6.3s -> 21ms

2) order, 20 iterations:
1M x 1K, sparsity=0.1:     13.4s -> 3.9s
100K x 10K, sparsity=0.1:  7.7s -> 510ms
10K x 100K, sparsity=0.1:  6.5s -> 64ms

Note that the improvements increase as the number of columns, or more
precisely, the number of non-zeros per row increase because there are
additional overheads (such as selection vector extraction or the actual
sorting) in the number of rows.