[HOTFIX][SYSTEMML-1659] Fix sum-sumSq aggregate elimination rewrite
This patch fixes the rewrite 'remove unnecessary aggregate' for the case
of sum-sumsq. As it turned out, the newly introduced tests, did not
check for it because sum(rowSums(X^2)) is first rewritten (w/ static
rewrites) to sum(X^2) and only subsequently rewritten (w/ dynamic
rewrites) to sumSq(X). For the NN library, the sum is exposed after
dynamic rewrites and hence the issue of the wrong aggregation type
assignment showed up.

This issue did not show up in local tests, because they were run with a
custom test suite, which only included application, conversion,
functions, and mlcontext tests (due to issues with the recently
introduced gpu tests).