[SYSTEMML-1683] Improved codegen row template (indexing, cbind)
This patch makes two improvements to the code generator row-wise
template in order to further reduce the number of intermediates in
scripts such as MLogreg as well as minor explain improvements):

(1) Column indexing support w/ unknown indexing expressions.

(2) Fusion of cbind with empty matrix after row-wise template (row
aggregates).

(3) Extended explain to show the line numbers (of the original script)
for generated operators.

For example, on MLogreg and a dense 100M x 10 scenario, this reduces the
buffer pool writes from (76/30/1) to (60/30/1) and execution time from
282s to 256s (compared to the baseline w/ existing fused operators of
529s). 

Note that there is substantial additional potential, which can be
exploited to reduce the number evictions. However, this will be
addressed in separate changes as it requires optimizer changes (e.g.,
considering multi-aggregates while considering materialization, and
materialization points per consumer).