[SYSTEMML-1684] New packed rmvar instructions w/ multiple operands
In SystemML, every scalar, frame, or matrix intermediate is eventually
removed from the symbol table (and potentially bufferpool and dfs) with
so-called 'remove variable' (rmvar) instructions. So far, these
instructions had the simple form 'rmvar _mvar1' which was unnecessarily
verbose regarding the corresponding EXPLAIN output (and partially
interpretation overhead). This patch adds a cleanup pass (after the
actual instruction generation) that collapses a sequence of rmvar
instructions into a single packed rmvar instruction with multiple
operands. The execution semantics of both representations are identical
but the plan is more concise. For example, for MLogreg, this patch
reduced the size of execution plans from 666/76 to 551/76 instructions
(w/ existing fused operators) and from 605/68 to 493/68 (w/ codegen
enabled).