[SYSTEMML-1686] Fix 2D transpose convolution filter shape & input size
Currently, the transpose conv2d layer (`nn/layers/conv2d_tranpose.dml`)
has a bug in which the filters tensor `W` has an incorrect shape, and
the `conv2d_backward_data` op has an incorrect input shape argument.
This results in an exception when the number of input channels `C` is
not equal to the number of filters `F` (i.e. number of output channels).
Since the transpose conv2d op is the gradient of the conv2d op, the
filter tensor needs to have the shape `(C, F, Hf, Wf)` for `F` filters,
rather than `(F, C, Hf, Wf)`, in order to map from an input with `C`
channels to an output with `F` channels during the input data gradient
function (`conv2d_backward_data`) that is used in the forward pass.
Additionally, the `input_shape` argument for `conv2d_backward_data`
needs to be `(N, F, Hout, Wout)`, rather than `(N, C, Hout, Wout)` in
order to map from an input with `C` channels to an output with `F`
channels.  Our previous test cases did not catch this issue because
the tests used `C = F = 1`.

The motivation here is to think about the shape of `W` during a regular
2D conv function, i.e. `(F, C, Hf, Wf)`.  In the forward *and* backward
pass of regular 2D conv, `W` maintains the same `(F, C, Hf, Wf)`shape.
Specifically, in the backward pass, the data gradient function
`conv2d_backward_data` accepts the same `W` of shape `(F, C, Hf, Wf)`
to map from gradients wrt the output of shape `(N, F, Hout, Wout)` to
gradients wrt to the input of shape `(N, C, Hin, Win)`.  In a 2D
transpose conv function, we use the 2D conv data gradient function as
the forward function.  If we use the same size terminology as from
regular 2D conv, we would be mapping from an input with `F` channels to
an output with `C` channels (which seems backward, but is correct since
the forward pass of 2D transpose conv is defined as the backward pass of
2D conv), and we would be using the same filter `W` of shape
`(F, C, Hf, Wf)`.  Now, in order to make the terminology of the 2D
transpose conv layer consistent with the rest of the library, we would
need to swap the variable names `F` and `C` so that we are mapping from
an input with `C` channels to an output with `F` channels.  Thus, we
would use a filter `W` of shape `(C, F, Hf, Wf`).

This commit fixes these issues, and improves the robustness of the
associated tests.

Closes #541.