[SYSTEMML-1691,1692] New IPA passes: literal replacement and rewrites
This patch introduces two new passes for inter-procedural analysis
(IPA): (1) literal propagation and replacement into functions, and (2)
static rewrites, which are both applied for any number of requires IPA
iterations. The internal abstraction for function call summaries has
been extended accordingly. The new literal propagation and replacement
works on a fine granularity of individual function parameters and
propagates any literals that are consistent across all function calls,
independent of remaining inputs. Together with the additional rewrites
pass, this allows rewrites such as constant folding and subsequent
removal of branches which can significantly cut down the program size
and number of distributed operations. 

For example, for GLM poisson.log, this change reduced the size of the
initial runtime program from 2132/153 to 1164/45 local/distributed
instructions, which is now much easier to debug and profile.