[SYSTEMML-1679] Add a new threshold utility function
This function accepts a matrix X and a threshold parameter thresh to get an indicator matrix with
values in {0, 1} depending on whether or not the values in X are above thresh. It can be used, for
example, to determine the predicted class in a binary classification problem given the output of
a sigmoid layer.

Closes #548.