[SYSTEMML-1716] Fix memory-efficiency spark rmm matrix mult operations
This patch improves the memory efficiency and performance of our spark
rmm matrix multiplication operator:

(1) Shallow block replication: Since we only need modified keys, we now
use shallow block replicas, which avoids (potentially a large number of)
unnecessary block copies before the join.

(2) Adjusted number of join partitions: The replication leads to
increased data sizes of intermediates. Using the number of input
partitions causes potentially very large partitions after the join which
creates memory pressure and limits the degree of parallelism. We now
also parallelize the matrix multiply over IJK blocks instead of IK
output blocks.

(3) Hash function: In order to enable IJK parallelization, we now use a
hash function over IJK instead of IK. Along with that, this patch also
replaces the historically used hash function for triple indexes with our
default hash function because the old function was known to have issues
with certain matrix shapes.

Finally, there is additional potential to remove the shuffle for final
aggregation (which is not addressed here). Down the road, we should use
to alternative rmm implementations - if the number of output blocks
already saturates the degree of parallelism and is known to have limited
memory requirements we should parallelize over IK and provide a custom
partitioner that can be carried through from the join to the final
aggregation (note the existing hadoop partitioner of triple indexes is
not used automatically).