[SYSTEMML-1718] Fix matrix export to non-default dfs file system
This patch fixes issues of the buffer pool export (which is also used
for persistent writes of in-memory matrices). In case the write file
scheme is different from the scratch space or input file scheme, an
export of binary matrices via copy fails because we cannot copy across
file systems.