[SYSTEMML-1675] Add a new 2D depthwise transpose convolution layer
This adds a new 2D depthwise transpose convolution layer.  A depthwise
transpose convolution (1) applies a different filter to each unique
group of M input channels separately, thus condensing each group of M
input channels to 1 output channel, and (2) concatenates the results
into a single volume with C/M output channels.  This is in contrast to
a regular 2D transpose convolution, in which all of the filters would be
applied to all of the input channels at once.

In addition to the new layer, this also adds the associated unit and
gradient tests.