[SYSTEMML-1674] Add a new 2D depthwise convolution layer
This adds a new 2D depthwise convolution layer.  A depthwise convolution
(1) applies a different set of M filters to each input channel
separately, thus expanding each input channel to M output channels, and
(2) concatenates the results into a single volume with C*M output
channels.  This is in contrast to a regular 2D convolution, in which all
of the filters would be applied to all of the input channels at once.

In addition to the new layer, this also adds the associated unit and
gradient tests.

Closes #542.