[SYSTEMML-1722] Fix mvvar to non-default file system (w/o rename)
This patch fixes the code path of mvvar to become aware of the source
and target file URI scheme in order to avoid a rename in case of
different schemes. 

Furthermore, this patch also removes the obsolete mv and rm file
instructions, which are unused for many years (since they were replaced
by mvvar and rmvar).