[SYSTEMML-1725] Extended codegen row-wise template (full rc aggregates)
This patch extends the code generator row-wise template by support for
full row/column aggregations (in addition to the types no_agg, row_agg,
col_agg). The motivation for this are scenarios where we currently
compile a sequence of row-wise and cell-wise templates, because the
patterns contain both intermediate row aggregates and a final full
aggregation. 

For example, on Mlogreg over an 100Mx10 input, this patch improved
end-to-end performance with codegen from 248s with (w/ 59/29/1
bufferpool writes) to 193s (w/ 45/15/1 bufferpool writes). In
comparison, without codegen but existing fused operators, the end-to-end
performance was 512s (w/ 234/30/1 writes).