[SYSTEMML-1727] Fix invalid mvvar instruction generation for pwrites
This patch fixes the generation of runtime instructions (piggybacking)
by removing the generation of mvvar instructions for persistent writes
that are the only consumer of a transient reads and write in format
binary block. Mvvar (move variable) instructions are invalid here for
two reasons. First, in case of different file URI schemes, the mvvar
cannot be realized as a rename (meta data operation) but reads the
matrix into CP, which leads to OOMs for large data. Second, in case of a
subsequent use of the variable, the temporary file does no longer exist,
which potentially leads to file not found exceptions. 

This patch fixes these issues by globally disabling the generation of
mvvar instructions for persistent writes. A more fine-grained handling
would be possible but would require a major rewrite of how we currently
generate instructions. This seems unnecessary, given that in CP and
Spark intermediates rarely reside on HDFS anyway.