[SYSTEMML-1717] Reduced alloc codegen row-wise vector intermediates
This patch improves the decision on the number of required thread-local
vector intermediates for codegen rowwise operations (similar in spirit
to register allocation in PL compilers). So far we used the number of
vector operations. Instead, we now identify unary vector pipelines which
allow the reuse of vector intermediates in the ring buffer of allocated
vectors. For long unary pipelines or larger vectors, this significantly
reduces the L1 data cache misses.