[SYSTEMML-1733] Fix correctness seq w/ small increment (num stability)
This patch fixes correctness issues (wrong output dimensions) of seq
with very small increment values in the compiler and runtime. The
original formulation of "1 + floor((to-from)/incr)" was prone to
round-off errors for very small from and increment values due to the
subtraction of values of very different magnitude. We now use a more
formulation that is more robust in that regard. This patch also
consolidates all compiler and runtime calls via a new primitive as well
as fixes a related issue of the recompiler.