[SYSTEMML-1678] Add a new 1D top_k utility function
This function computes the top k values (i.e. probabilities) and
associated indices (i.e. classes) from the input matrix X.  A typical
use case is that in which X is the output of a softmax layer, and
values and indices will contain rows with the top k probabilities and
class indices.

Closes #551.