[SYSTEMML-1736] Add a new 2D top_k utility function
This function computes the top k values (i.e. probabilities) and
associated indices (i.e. classes) from the input matrix X. A typical use
case here is that in which X is the output of a 2D softmax layer, so
each channel contains a set of normalized class probalilities, and
values and indices will contain the top k probabilities and indices
along the channel axis. The scenario will be common in an image
segmentation problem.

Closes #551.