[SYSTEMML-1676] Add a new 2D softmax layer to the `nn` library
A 2D softmax layer would accept a tensor of shape (N,C,H,W), where the
C axis contains scores for D classes, and output a tensor of the same
shape, with the scores transformed to normalized probabilities. The
typical use case would be a segmentation problem, in which every pixel
has a multiclass prediction.

Closes #555.