[MINOR] Fix mlcontext missing explain header and statistics reset
This patch fixes two minor issues of the new mlcontext. First, it adds
the missing explain header of available memory budgets and degree of
parallelism by calling a new primitive that prepares the explain output
(called for consistency from dmlscript and mlcontext). This also
includes a slight modification of runtime explain which now avoids an
unnecessary pass over the program. Second, it adds reset calls for the
number of executed jobs and cache statistics which would otherwise
accumulate over multiple runs of dml scripts in the same process.