[SYSTEMML-1747] New rewrite and performance binary outer operations
This patch adds a new rewrite to avoid unnecessary column replication
for subsequent binary matrix-vector operations, which can be directly
expressed via outer: (a %*% ones) == b) -> outer(a, b, "=="). As a
result of this rewrite, there are fewer intermediates and smaller memory
requirements (2*matrix + vector -> matrix + 2*vector).

Furthermore, this patch also makes a couple of performance improvements
in the related outer operations, specifically (1) shallow instead of
deep dense vector copies, and (2) avoid binary search for outer
comparison operations with sorted rhs vector if the number of columns is
very small.