[SYSTEMML-1767] Performance codegen rowwise template w/ column agg
This patch makes the codegen row-wise template consistent with the
mmchain operation in terms of its condition to fallback to
single-threaded operations if the temporary memory for partial
aggregations exceed the internal threshold. On a scenario with 2M spase
features, this patch improved performance by 20x because it avoids
unnecessary L3 cache thrashing.