[SYSTEMML-1762] Fix spark/mr reshape instructions (missing blocks)
This patch fixes special cases of spark/mr reshape instructions, which
applies to sparse and dense. Given an input block, we compute the first
and last output block index and create all blocks in between. In cases
where we output at least three blocks and the first and last block have
the same row index (e.g., (1,12) and (1,1)), we missed to create the
remaining blocks after the first till the end of the row block (e.g.,
the blocks (1,13) and (1,14) if the matrix has 14 column blocks),
leading to null pointer exceptions on the actual reshape copy. 

Furthermore, this patch also cleans up the respective sparse and dense
reshape interfaces and implementations as well as adds tests for the
encountered special cases.